+++ tp3/entregable/tp3-bundle.v1/src/sched.h
@@ -14,7 +14,7 @@
 #include "tss.h"
 
 unsigned short sched_proximo_indice();
-void start_zombie(u8 player, u8 class, u8 y);
+void start_zombie(u8 player, u8 class, u32 y);
 
 
 #endif	/* !__SCHED_H__ */
